# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  admin.ChannelResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      wallet_id:
        type: integer
    type: object
  admin.CreateChannelRequest:
    properties:
      name:
        type: string
      wallet_id:
        type: integer
    required:
    - name
    - wallet_id
    type: object
  admin.CreateWalletRequest:
    properties:
      balance:
        type: number
    type: object
  admin.TopUpWalletRequest:
    properties:
      amount:
        type: number
      reference_id:
        type: string
    required:
    - amount
    type: object
  admin.TransactionResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      reference_id:
        type: string
      type:
        type: string
      wallet_id:
        type: integer
    type: object
  admin.WalletResponse:
    properties:
      balance:
        type: number
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  handler.SwaggerDeliveryReport:
    properties:
      attempts:
        type: integer
      channel:
        type: string
      created_at:
        type: string
      delivered_at:
        type: string
      destination:
        type: string
      last_error:
        type: string
      line:
        type: string
      message:
        type: string
      message_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      wallet_id:
        type: integer
    type: object
  handler.SwaggerErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  handler.SwaggerSMSRequest:
    properties:
      channel:
        example: main
        type: string
      dest:
        example: "989121234567"
        type: string
      line:
        example: express
        type: string
      message:
        example: Your verification code is 123456
        type: string
    type: object
  handler.SwaggerSMSResponse:
    properties:
      accepted:
        type: boolean
      channel:
        type: string
      cost:
        type: number
      dest:
        type: string
      line:
        type: string
      message:
        type: string
      message_id:
        type: string
      remaining_balance:
        type: number
    type: object
info:
  contact: {}
  description: Asynchronous SMS gateway with Redis credit reservations and delivery
    reports.
  title: SMS Gateway API
  version: "1.0"
paths:
  /channels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin.ChannelResponse'
            type: array
      security:
      - AdminAPIKey: []
      summary: List SMS channels
      tags:
      - Channels
    post:
      consumes:
      - application/json
      parameters:
      - description: Channel request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin.ChannelResponse'
      security:
      - AdminAPIKey: []
      summary: Create an SMS channel
      tags:
      - Channels
  /sms:
    post:
      consumes:
      - application/json
      description: Reserves credit and queues an asynchronous SMS delivery.
      parameters:
      - description: SMS request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SwaggerSMSRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handler.SwaggerSMSResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.SwaggerErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.SwaggerErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.SwaggerErrorResponse'
      summary: Queue an SMS
      tags:
      - SMS
  /sms/{message_id}:
    get:
      parameters:
      - description: SMS message ID
        in: path
        name: message_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SwaggerDeliveryReport'
      security:
      - AdminAPIKey: []
      summary: Get an SMS delivery report
      tags:
      - SMS Reports
  /wallets:
    post:
      consumes:
      - application/json
      parameters:
      - description: Wallet request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.CreateWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin.WalletResponse'
      security:
      - AdminAPIKey: []
      summary: Create a wallet
      tags:
      - Wallets
  /wallets/{id}:
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.WalletResponse'
      security:
      - AdminAPIKey: []
      summary: Get a wallet balance
      tags:
      - Wallets
  /wallets/{id}/sms:
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - default: 100
        description: Maximum reports (1-500)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.SwaggerDeliveryReport'
            type: array
      security:
      - AdminAPIKey: []
      summary: List SMS delivery reports for a wallet
      tags:
      - SMS Reports
  /wallets/{id}/topup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      - description: Top-up request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.TopUpWalletRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.WalletResponse'
      security:
      - AdminAPIKey: []
      summary: Top up a wallet
      tags:
      - Wallets
  /wallets/{id}/transactions:
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin.TransactionResponse'
            type: array
      security:
      - AdminAPIKey: []
      summary: List wallet transactions
      tags:
      - Wallets
securityDefinitions:
  AdminAPIKey:
    in: header
    name: X-Admin-API-Key
    type: apiKey
swagger: "2.0"